// src/app/modules/character/models/character.model.ts
export interface ResponseCharacterModel {
    count: number;
    next: string | null;
    previous: string | null;
    results: Array<Character>;
}

export interface Character {
    name: string;
    height: string;
    mass: string;
    hair_color: string;
    skin_color: string;
    eye_color: string;
    birth_year: string;
    gender: string;
    homeworld: string;
    films: Array<string>;
    species: Array<string>;
    vehicles: Array<string>;
    starships: Array<string>;
    created: string;
    edited: string;
    url: string;
}

export interface CharacterSortOption {
    title: string;
    type: string;
    sort: string;
}